refactor: Repository 모두 제거, Service로 위임
- 테스트에도 반영
