+++ src/test/java/subway/application/SubwayServiceTest.java
@@ -4,8 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.doReturn;
 import static subway.TestFixture.LINE_A;
 import static subway.TestFixture.LINE_B;
-import static subway.TestFixture.SHORTEST_PATH_IN_LINE_A_AND_B_STATION_A_TO_E;
-import static subway.TestFixture.SHORTEST_PATH_IN_LINE_C_AND_D_STATION_A_TO_E;
 import static subway.TestFixture.SHORTEST_PATH_STATIONS_IN_LINE_A_AND_B_STATION_A_TO_E;
 import static subway.TestFixture.STATION_A;
 import static subway.TestFixture.STATION_E;
@@ -20,9 +18,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import subway.domain.Path;
 import subway.domain.Station;
-import subway.persistence.LineRepository;
 import subway.persistence.StationDao;
 
 @ExtendWith(MockitoExtension.class)
@@ -32,7 +28,7 @@ class SubwayServiceTest {
     private SubwayService subwayService;
 
     @Mock
-    private LineRepository lineRepository;
+    private LineService lineService;
     @Mock
     private StationDao stationDao;
 
@@ -42,7 +38,7 @@ class SubwayServiceTest {
         doReturn(List.of(
                 LINE_A,
                 LINE_B
-        )).when(lineRepository).findAll();
+        )).when(lineService).findAll();
         doReturn(STATION_A).when(stationDao).findById(STATION_A.getId());
         doReturn(STATION_E).when(stationDao).findById(STATION_E.getId());
         var expectedIds = SHORTEST_PATH_STATIONS_IN_LINE_A_AND_B_STATION_A_TO_E.stream()
@@ -50,7 +46,6 @@ class SubwayServiceTest {
                 .collect(Collectors.toList());
 
         var path = subwayService.getShortestPath(STATION_A.getId(), STATION_E.getId());
-
 
         assertThat(path.getStations())
                 .extracting("id")
